\name{Medbetareg-package}
\alias{Medbetareg-package}
\alias{Medbetareg}
\docType{package}
\title{Generalized Beta Regression to Elicit Conditional Distributions of Medical Variables}
\description{Software implementing the methodology proposed by Magrini \emph{et al.} (2018).}
\details{
  \tabular{ll}{
  Package: \tab Medbetareg\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2018-06-25\cr
  License: \tab GPL-2\cr
  }
  
The elicitation of an informative prior distribution on parameters of
a univariate conditional model is typically difficult when the response
is a medical continuous variable, because some level of statistical
training is required to a medical expert for interpreting parameters
and for retrieving appropriate quantitative information about them.
Magrini \emph{et al.} (2018) proposed a Generalized Beta regression where
parameter elicitation is performed by establishing a correspondence
among measured values expressed as relative positions within intervals
with a clinical interpretation, regardless the original scales of variables.}
\author{Alessandro Magrini <alessandro.magrini@unifi.it>}
\references{A. Magrini, D. Luciani, and F. M. Stefanini (2018).
  "Generalized Beta regression to elicit conditional distributions of medical variables".
  \emph{Austrian Journal of Statistics}, 47(3): 20-38. DOI: 10.17713/ajs.v47i3.629.}
